+++ src/guestbooks/domain/useCases/downloadGuestbookResponsesByCollectionId.ts
@@ -1,0 +1,8 @@
+import { GuestbookRepository } from '../repositories/GuestbookRepository'
+
+export function downloadGuestbookResponsesByCollectionId(
+  guestbookRepository: GuestbookRepository,
+  collectionIdOrAlias: number | string
+): Promise<string> {
+  return guestbookRepository.downloadGuestbookResponsesByCollectionId(collectionIdOrAlias)
+}