+++ src/guestbooks/domain/useCases/downloadGuestbookResponsesByDataverseId.ts
@@ -1,8 +1,0 @@
-import { GuestbookRepository } from '../repositories/GuestbookRepository'
-
-export function downloadGuestbookResponsesByDataverseId(
-  guestbookRepository: GuestbookRepository,
-  dataverseId: number | string
-): Promise<string> {
-  return guestbookRepository.downloadGuestbookResponsesByDataverseId(dataverseId)
-}