+++ tests/component/sections/dataset/dataset-guestbook/DatasetGuestbook.spec.tsx
@@ -27,8 +27,9 @@ describe('DatasetGuestbook', () => {
       createGuestbook: cy.stub(),
       getGuestbook: cy.stub(),
       getGuestbooksByCollectionId: cy.stub(),
+      getGuestbookResponsesByGuestbookId: cy.stub(),
       setGuestbookEnabled: cy.stub(),
-      downloadGuestbookResponsesByDataverseId: cy.stub(),
+      downloadGuestbookResponsesByCollectionId: cy.stub(),
       downloadGuestbookResponsesOfAGuestbook: cy.stub(),
       assignDatasetGuestbook: cy.stub().resolves(undefined),
       removeDatasetGuestbook: cy.stub().resolves(undefined)