+++ tests/component/sections/edit-dataset-terms/EditDatasetTerms.spec.tsx
@@ -107,8 +107,9 @@ describe('EditDatasetTerms', () => {
       createGuestbook: cy.stub(),
       getGuestbook: cy.stub(),
       getGuestbooksByCollectionId: cy.stub().resolves(mockGuestbooks),
+      getGuestbookResponsesByGuestbookId: cy.stub(),
       setGuestbookEnabled: cy.stub(),
-      downloadGuestbookResponsesByDataverseId: cy.stub(),
+      downloadGuestbookResponsesByCollectionId: cy.stub(),
       downloadGuestbookResponsesOfAGuestbook: cy.stub(),
       assignDatasetGuestbook: cy.stub().resolves(undefined),
       removeDatasetGuestbook: cy.stub().resolves(undefined)