+++ tests/component/sections/edit-dataset-terms/useAssignDatasetGuestbook.spec.tsx
@@ -13,8 +13,9 @@ describe('useAssignDatasetGuestbook', () => {
       createGuestbook: cy.stub(),
       getGuestbook: cy.stub(),
       getGuestbooksByCollectionId: cy.stub(),
+      getGuestbookResponsesByGuestbookId: cy.stub(),
       setGuestbookEnabled: cy.stub(),
-      downloadGuestbookResponsesByDataverseId: cy.stub(),
+      downloadGuestbookResponsesByCollectionId: cy.stub(),
       downloadGuestbookResponsesOfAGuestbook: cy.stub(),
       assignDatasetGuestbook: cy.stub(),
       removeDatasetGuestbook: cy.stub()